Enhanced Provider Dashboard & Vehicle Management

+++ provider/dashboard.php
@@ -11,6 +11,9 @@ if (!is_provider_logged_in()) {
 
 $provider_id = $_SESSION['provider_id'];
 
+// Get optional success message
+$msg = isset($_GET['msg']) ? htmlspecialchars($_GET['msg']) : '';
+
 // Database connection
 $conn = Connect();
 
@@ -56,11 +59,24 @@ $totalEarnings = $result->fetch_assoc()['total'];
     <link href="../style.css" rel="stylesheet">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
     <script src="https://kit.fontawesome.com/yourkitid.js" crossorigin="anonymous"></script>
+    <script>
+        // Confirm deletion popup
+        function confirmDelete(vehicleName) {
+            return confirm("Are you sure you want to delete " + vehicleName + "?");
+        }
+    </script>
 </head>
 <body>
     <?php include __DIR__ . '/../includes/provider_navbar.php'; ?>
 
     <div class="container main-content">
+        <!-- Success message -->
+        <?php if(!empty($msg)): ?>
+        <div class="alert alert-success mt-3">
+            <?php echo $msg; ?>
+        </div>
+        <?php endif; ?>
+
         <!-- Stats Grid -->
         <div class="row mb-4">
             <div class="col-md-3">
@@ -115,8 +131,8 @@ $totalEarnings = $result->fetch_assoc()['total'];
                             <td><span class='badge bg-".($vehicle['vehicle_availability']=='yes'?'success':'warning')."'>".ucfirst($vehicle['vehicle_availability'])."</span></td>
                             <td>\${$vehicle['price']}</td>
                             <td>
-                                <a href='edit_vehicle.php?id={$vehicle['vehicle_id']}' class='btn btn-sm btn-light'><i class='fas fa-edit'></i></a>
-                                <a href='delete_vehicle.php?id={$vehicle['vehicle_id']}' class='btn btn-sm btn-danger'><i class='fas fa-trash'></i></a>
+                                <a href='edit_vehicle.php?id={$vehicle['vehicle_id']}' class='btn btn-sm btn-light'>Edit<i class='fas fa-edit'></i></a>
+                                <a href='delete_vehicle.php?id={$vehicle['vehicle_id']}' class='btn btn-sm btn-danger' onclick='return confirmDelete(\"{$vehicle['vehicle_model']}\");'>Delete<i class='fas fa-trash'></i></a>
                             </td>
                         </tr>";
                     }
